Rename pokemons table to pokemons_assets
Prior to this change, the DB table which stores the pokemons assets was named pokemons which might give the wrong impression about what the table stores. The table stores the pokemon gif and image file urls. In this change we change the fetch calls to reflect what the table stores and distinguish it from the pokemons api.