Update teams endpoint to use fed_teams table
Prior to this change, we used the team_members table to figure out the unique teams. Now that we created a fed_teams table, we can use it to get all the unique team names rather than use the team_members table where we have to filter the entries to get the unique teams.